Update duplicate_message_policing.py
Please add the following feature:
Currently, the policing feature doesn't handle messages to the same case, e.g. lol and Lol are both distinct. This needs to end. However, the database includes hashes of old which will be different for lol and Lol as an example. So, we need to keep backwards compatibility, but all new hashes should be of the .lower() message content. So we will need to try two inserts in the try: block, basically.